Finally get something rendered on the screen, still working on model import system

+++ VoxelEngineVulkan/src/GraphicAPI/Vulkan/VulkanClasses.cpp
@@ -1896,7 +1896,7 @@ void vulkan::VulkanResouceManager::ConstructVulkanRenderObject(
 	for (auto& x : modeldata.meshes) {
 		_BufferManager.CreateVertexBuffer1(x.vertices, renderObject.vertexBuffer, renderObject.vertexmemory);
 		_BufferManager.CreateIndexBuffer1(x.indices, renderObject.indiceBuffer, renderObject.indicememory);
-		renderObject.indiceCounts.push_back(x.indices.size());
+		renderObject.indiceCounts.push_back(static_cast<uint32_t>(x.indices.size()));
 	}
 
 	renderObject.pipeline = pipeline;
