+++ VoxelEngineVulkan/src/GraphicAPI/Vulkan/VulkanCommon.h
@@ -119,13 +119,13 @@ namespace vulkan {
 		//Non-const Function
 		//该函数用于将各个vector中数据的地址赋值给各个创建信息中相应成员，并相应改变各个count
 		void UpdateAllArrays() {
-			createInfo.stageCount = shaderStages.size();
-			vertexInputStateCi.vertexBindingDescriptionCount = vertexInputBindings.size();
-			vertexInputStateCi.vertexAttributeDescriptionCount = vertexInputAttributes.size();
-			viewportStateCi.viewportCount = viewports.size() ? uint32_t(viewports.size()) : dynamicViewportCount;
-			viewportStateCi.scissorCount = scissors.size() ? uint32_t(scissors.size()) : dynamicScissorCount;
-			colorBlendStateCi.attachmentCount = colorBlendAttachmentStates.size();
-			dynamicStateCi.dynamicStateCount = dynamicStates.size();
+			createInfo.stageCount = static_cast<uint32_t>(shaderStages.size());
+			vertexInputStateCi.vertexBindingDescriptionCount = static_cast<uint32_t>(vertexInputBindings.size());
+			vertexInputStateCi.vertexAttributeDescriptionCount = static_cast<uint32_t>(vertexInputAttributes.size());
+			viewportStateCi.viewportCount = uint32_t(viewports.size()) ? uint32_t(viewports.size()) : dynamicViewportCount;
+			viewportStateCi.scissorCount = uint32_t(scissors.size()) ? uint32_t(scissors.size()) : dynamicScissorCount;
+			colorBlendStateCi.attachmentCount = static_cast<uint32_t>(colorBlendAttachmentStates.size());
+			dynamicStateCi.dynamicStateCount = static_cast<uint32_t>(dynamicStates.size());
 			UpdateAllArrayAddresses();
 		}
 	private:
@@ -176,7 +176,7 @@ namespace vulkan {
 			return true;
 		}
 		void UpdateAllArray() {
-			bindingCounts = bindings.size();
+			bindingCounts = static_cast<uint32_t>((bindings.size()));
 		}
 	};
 
@@ -311,6 +311,7 @@ namespace vulkan {
 		uint32_t height{ 0 };
 		uint32_t mipLevels{ 0 };
 	};
+
 	struct VulkanRenderObject {
 		std::string name;
 		VkPipeline pipeline;
